% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{colour}
\alias{colour}
\title{Colour text for terminal output}
\usage{
colour(strings, col = "default", mod = c("none", "bold", "dim", "italic"))
}
\arguments{
\item{strings}{A character vector to colour.}

\item{col}{A string naming the colour to use: one of \code{"black"},
\code{"red"}, \code{"green"}, \code{"yellow"}, \code{"blue"}, \code{"magenta"}, \code{"cyan"},
\code{"white"} or \code{"default"}.}

\item{mod}{A string giving a modifier to apply, one of \code{"none"},
\code{"bold"}, \code{"dim"} or \code{"italic"}.}
}
\value{
A character vector of the same length as \code{strings}, with ANSI
escape codes added (unless colour output is disabled).
}
\description{
This function wraps strings in ANSI escape codes so that they are shown
in colour (and optionally bold, dim or italic) when printed to a
capable terminal. If the \code{TRACTOR_NOCOLOUR} environment variable is set
to a true value, the strings are returned unmodified.
}
\author{
Jon Clayden
}
