% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{experimentPaths}
\alias{experimentPaths}
\title{Standard search paths for experiment scripts}
\usage{
experimentPaths()
}
\value{
A character vector of directory paths.
}
\description{
This function returns the ordered list of directories that are searched
for TractoR experiment scripts, by \code{\link[=findExperiment]{findExperiment()}} and
\code{\link[=scanExperiments]{scanExperiments()}}. In order, these are the current working directory,
the user's \verb{~/.tractor} directory, any directories listed in the
\code{TRACTOR_PATH} environment variable, the \code{tractor/experiments}
directories of any packages named in the \code{TRACTOR_PACKAGES} environment
variable, and finally the standard experiment directory shipped with
TractoR.
}
\author{
Jon Clayden
}
