% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{findExperiment}
\alias{findExperiment}
\title{Locate an experiment script by name}
\usage{
findExperiment(exptName)
}
\arguments{
\item{exptName}{A string giving the name of the experiment, without the
\code{.R} file extension.}
}
\value{
A string giving the path to the matching experiment script.
}
\description{
This function searches the standard experiment script search paths (see
\code{\link[=experimentPaths]{experimentPaths()}}) for a script with the given name, and returns the
path to the first match. An error is raised if no matching script can be
found.
}
\author{
Jon Clayden
}
