% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{nArguments}
\alias{nArguments}
\title{The number of arguments passed to an experiment script}
\usage{
nArguments()
}
\value{
An integer giving the number of arguments, or zero if no
\code{Arguments} vector currently exists.
}
\description{
This function returns the number of command-line-style arguments that
were passed to the currently running experiment script, as stored in
the global \code{Arguments} vector by \code{\link[=bootstrapExperiment]{bootstrapExperiment()}}.
}
\author{
Jon Clayden
}
