% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parallelApply}
\alias{parallelApply}
\title{Apply a function over a vector or list, in parallel if possible}
\usage{
parallelApply(x, fun, ..., preschedule = TRUE, setSeed = TRUE,
  silent = FALSE, cores = NULL)
}
\arguments{
\item{x}{A vector or list to iterate over.}

\item{fun}{A function to apply to each element of \code{x}.}

\item{...}{Additional arguments to \code{fun}.}

\item{preschedule}{Boolean value passed as \code{mc.preschedule} to
\code{parallel::mclapply()}.}

\item{setSeed}{Boolean value passed as \code{mc.set.seed} to
\code{parallel::mclapply()}.}

\item{silent}{Boolean value passed as \code{mc.silent} to
\code{parallel::mclapply()}.}

\item{cores}{An integer giving the number of cores to use. If \code{NULL},
the default, the \code{mc.cores} or \code{cores} option is used, in that order
of preference, falling back to two if neither is set.}
}
\value{
A list of results, one per element of \code{x}, as for \code{\link[=lapply]{lapply()}}.
}
\description{
This function applies \code{fun} to each element of \code{x} in turn, using
\code{parallel::mclapply()} if the \code{parallel} package is attached, or falling
back to a plain \code{\link[=lapply]{lapply()}} call otherwise. When running in parallel, the
\code{reportr} message prefix is temporarily adjusted to include the process
ID, so that messages from concurrent workers can be told apart.
}
\author{
Jon Clayden
}
