% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{requireArguments}
\alias{requireArguments}
\title{Require that an experiment script was given sufficient arguments}
\usage{
requireArguments(..., name = TRUE)
}
\arguments{
\item{...}{Either a single integer giving the minimum number of
arguments required, or one or more strings giving names for the
expected arguments (in which case their number also becomes the
minimum required).}

\item{name}{Boolean value: if \code{TRUE}, and names were given via \code{...},
the global \code{Arguments} vector is renamed accordingly. This can cause
problems in some contexts, such as when \code{Arguments} is used with
\code{apply}-family functions, in which case \code{FALSE} should be used.}
}
\value{
Called for its side effect of checking (and, optionally, naming)
the arguments. \code{NULL} is returned invisibly.
}
\description{
This function checks that the currently running experiment script was
passed a sufficient number of command-line-style arguments (see
\code{\link[=nArguments]{nArguments()}}), raising an error if not. If names are given, the
global \code{Arguments} vector is also named accordingly, so that the
expected arguments can subsequently be accessed by name as well as by
position.
}
\author{
Jon Clayden
}
