% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{readYaml}
\alias{readYaml}
\alias{writeYaml}
\title{Read and write configuration variables in YAML format}
\usage{
readYaml(fileNames)

writeYaml(object, fileName)
}
\arguments{
\item{fileNames}{A character vector, or single colon-separated string,
of paths to YAML files. Elements which are empty or contain no
non-whitespace characters are ignored. If more than one file is given,
the resulting lists are merged, with later files' values taking
priority for duplicate keys (see \code{deduplicate()}).}

\item{object}{A list of configuration variables to write out.}

\item{fileName}{A string giving the path to write to.}
}
\value{
For \code{readYaml()}, a list of configuration variables read from
the file(s). \code{writeYaml()} is called for its side effect.
}
\description{
These functions read and write TractoR configuration variables using the
YAML format, via the \code{yaml} package. \code{readYaml()} is tolerant of a
missing space after the colon in \code{key:value} pairs, which is a common
shorthand used when writing configuration on the command line.
}
\author{
Jon Clayden
}
